backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
motor==3.6.0
pymongo==4.9.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pydantic-settings==2.6.1
httpx==0.28.1
python-dotenv==1.0.1
scikit-learn==1.5.2; python_version < "3.14"
pandas==2.2.3
numpy==2.1.3
joblib==1.4.2
pytest==8.3.4
pytest-asyncio==0.24.0
email-validator==2.2.0
